fix(ci): 修复MusicService和PlayerScreen剩余编译错误
- 移除MusicService.kt中不存在的androidx.media.app.Notification.MediaStyle
- 修复PlayerScreen.kt中onSeekTo参数类型从toInt改为toLong

+++ app/src/main/java/app/music_q36_qwen_code/ui/screens/PlayerScreen.kt
@@ -461,7 +461,7 @@ fun ProgressBar(
             value = sliderPosition,
             onValueChange = { sliderPosition = it },
             onValueChangeFinished = {
-                onSeekTo((sliderPosition * duration).toInt())
+                onSeekTo((sliderPosition * duration).toLong())
             },
             colors = SliderDefaults.colors(
                 thumbColor = Color(0xFF81C784),
